requirements.txt
flask
transformers
torch
sentence-transformers
faiss-cpu
PyPDF2
flask-cors